(doc): VPN guide connect with Vagrant+Virtualbox
There was a situation where a user using a MacOS device couldn't connect
to the VPN with the 'Global Protection' app and there wasn't other
alternatives.

It is importante to have a simple guide on how to use Vagrant with Virtualbox
to connect to the VPN as a last resource for users of Windows and MacOS devices.